@@ -1,9 +1,10 @@
 package com.example.floatingnotification.animator;
 
-import android.support.v4.view.ViewCompat;
-import android.support.v4.view.ViewPropertyAnimatorCompat;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import androidx.core.view.ViewCompat;
+import androidx.core.view.ViewPropertyAnimatorCompat;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class SlideInOutRightItemAnimator extends BaseItemAnimator {
 
