@@ -1,10 +1,11 @@
 package com.example.floatingnotification.viewholders;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.floatingnotification.R;
 import com.example.floatingnotification.models.NotificationItem;
